Rename project directories
Originally Project/ was meant as a placeholder name, and the user would copy and rename it before using. Later, Pheast/ was added as a second stage, which could lead to confusion as to why only one of them was named Project. `phenotyping/` makes it more clear that that is for the first, phenotyping stage.

Also, since users often run their data directly in the repo, the new directory names are lowercase for typing convenience.