Update phenotype ID format
Now all grouped modalities have a double underscore separating the gene ID from the other info, instead of colon (which causes issues when used in file names on some systems) or period (which is difficult to parse if gene IDs contain a period).

Additional colon and period separators in the IDs have been replaced with single underscores.

Fixes #11