Check for Ensembl GTF
Currently only GTFs from Ensembl are compatible with Pantry. The snakemake code now checks the first 1000 lines of the GTF file to ensure it is from Ensembl to prevent the user from running much of the pipeline and then encountering the compatibility issue. Fixes #12 followup issue in comments (though a full fix will require change to make more GTFs compatible).