New exonic/intronic region definitions
Instead of reporting exons found in at least a certain percentage of isoforms, report regions that are covered by exons in at least a certain percentage of isoforms. That allows, e.g. a region to be included if some isoforms have exons that extend beyond the region.

Intronic regions are defined as regions within the gene (as defined by exons) that contain no exons in any isoform.