Fix TWAS weights list order, etc.
The `*.profile` output file has the phenotype ID in the first column and has the same order as the input file list. The files in the input list are named as {phenotype_id}.wgt.RDat. The rows should be sorted by phenotype ID, but the input file list was sorted by the full filename, which can lead to a slightly different order, e.g. for phenotype IDs and 'Abc', 'Abcd'. The input is now sorted according to only the phenotype_id portion of the name.

`set -e` was removed from the FUSION weights script due to occasional errors related to models training in otherwise valid runs.